+++ src/main/java/com/saituo/order/entity/user/ProductOrderHis.java
@@ -3,9 +3,9 @@ package com.saituo.order.entity.user;
 public class ProductOrderHis {
 	
 	//序号
-	private Long  id;
+	private Integer  id;
 	//产品订单编码
-	private Long  registerNumber;
+	private Integer  registerNumber;
 	//类型结果：1.审批状态 2产品订单项状态 3发票状态
 	private String orderResult;
 	//创建者
@@ -13,10 +13,10 @@ public class ProductOrderHis {
 	//创建时间
 	private String acceptDate;
 
-	public Long getRegisterNumber() {
+	public Integer getRegisterNumber() {
 		return registerNumber;
 	}
-	public void setRegisterNumber(Long registerNumber) {
+	public void setRegisterNumber(Integer registerNumber) {
 		this.registerNumber = registerNumber;
 	}
 	public String getOrderResult() {
@@ -37,10 +37,10 @@ public class ProductOrderHis {
 	public void setAcceptDate(String acceptDate) {
 		this.acceptDate = acceptDate;
 	}
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 }
